rename pre-commit scripts
